+++ resources/views/contract/index.blade.php
@@ -81,21 +81,27 @@
                                                 </div>
                                             @endif
                                         @endif
+                                        @can('show contract')
                                             <div class="action-btn bg-warning ms-2">
                                                 <a href="{{ route('contract.show',$contract->id) }}"
                                                    class="mx-3 btn btn-sm d-inline-flex align-items-center"
                                                    data-bs-whatever="{{__('View Budget Planner')}}" data-bs-toggle="tooltip"
                                                    data-bs-original-title="{{__('View')}}"> <span class="text-white"> <i class="ti ti-eye"></i></span></a>
                                             </div>
+                                        @endcan
+                                        @can('edit contract')
                                             <div class="action-btn bg-info ms-2">
                                                 <a href="#" class="mx-3 btn btn-sm d-inline-flex align-items-center" data-url="{{ route('contract.edit',$contract->id) }}" data-ajax-popup="true" data-size="md" data-bs-toggle="tooltip" title="{{__('Edit')}}" data-title="{{__('Edit Contract')}}">
                                                     <i class="ti ti-pencil text-white"></i>
                                                 </a></div>
+                                        @endcan
+                                        @can('delete contract')
                                             <div class="action-btn bg-danger ms-2">
                                                 {!! Form::open(['method' => 'DELETE', 'route' => ['contract.destroy', $contract->id]]) !!}
                                                 <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}"><i class="ti ti-trash text-white"></i></a>
                                                 {!! Form::close() !!}
                                             </div>
+                                        @endcan
                                     </td>
                                 </tr>
                             @endforeach
