+++ resources/views/document/index.blade.php
@@ -10,9 +10,12 @@
 
 @section('action-btn')
     <div class="float-end">
+        @can('create document type')
             <a href="#" data-url="{{ route('document.create') }}" data-ajax-popup="true" data-title="{{__('Create New Document')}}" data-bs-toggle="tooltip" title="{{__('Create')}}"  class="btn btn-sm btn-primary">
                 <i class="ti ti-plus"></i>
             </a>
+
+        @endcan
     </div>
 @endsection
 
@@ -51,16 +54,21 @@
 
                                     @if(Gate::check('edit document type') || Gate::check('delete document type'))
                                         <td>
+                                            @can('edit document type')
                                                 <div class="action-btn bg-primary ms-2">
                                                     <a href="#" class="mx-3 btn btn-sm align-items-center" data-url="{{ URL::to('document/'.$document->id.'/edit') }}" data-ajax-popup="true" data-title="{{__('Edit Document Type')}}" data-bs-toggle="tooltip" title="{{__('Edit')}}" data-original-title="{{__('Edit')}}">
                                                         <i class="ti ti-pencil text-white"></i>
                                                     </a>
                                                 </div>
+                                            @endcan
+
+                                            @can('delete document type')
                                                 <div class="action-btn bg-danger ms-2">
                                                     {!! Form::open(['method' => 'DELETE', 'route' => ['document.destroy', $document->id],'id'=>'delete-form-'.$document->id]) !!}
                                                     <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}"><i class="ti ti-trash text-white text-white"></i></a>
                                                     {!! Form::close() !!}
                                                 </div>
+                                            @endcan
 
                                         </td>
                                     @endif
