+++ resources/views/jobStage/index.blade.php
@@ -41,9 +41,12 @@
 
 @section('action-btn')
     <div class="float-end">
+        @can('create job stage')
             <a href="#" data-url="{{ route('job-stage.create') }}" data-ajax-popup="true" data-title="{{__('Create New Job Stage')}}" data-bs-toggle="tooltip" title="{{__('Create')}}"  class="btn btn-sm btn-primary">
                 <i class="ti ti-plus"></i>
             </a>
+
+        @endcan
     </div>
 @endsection
 
@@ -65,16 +68,20 @@
                                             <span>{{$stage->title}}</span>
                                         </h6>
                                         <span class="float-end">
+                                            @can('edit job stage')
                                                 <div class="action-btn bg-primary ms-2">
                                                     <a href="#" class="mx-3 btn btn-sm align-items-center" data-url="{{ route('job-stage.edit',$stage->id) }}" data-ajax-popup="true" data-title="{{__('Edit Job Stage')}}" data-bs-toggle="tooltip" title="{{__('Edit')}}" data-original-title="{{__('Edit')}}">
                                                         <i class="ti ti-pencil text-white"></i>
                                                     </a>
                                                 </div>
+                                            @endcan
+                                            @can('delete job stage')
                                                     <div class="action-btn bg-danger ms-2">
                                                         {!! Form::open(['method' => 'DELETE', 'route' => ['job-stage.destroy', $stage->id],'id'=>'delete-form-'.$stage->id]) !!}
                                                         <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}"><i class="ti ti-trash text-white text-white"></i></a>
                                                         {!! Form::close() !!}
                                                      </div>
+                                                    @endcan
                                             </span>
                                     </li>
                                 @endforeach
