+++ resources/views/meeting/calender.blade.php
@@ -16,6 +16,7 @@
 
 
 @section('action-btn')
+    @can('create meeting')
         <div class="float-end">
             <a href="{{ route('meeting.index') }}" class="btn btn-sm btn-primary" data-bs-toggle="tooltip" title="{{__('List View')}}" data-original-title="{{__('List View')}}">
                 <i class="ti ti-list"></i>
@@ -24,6 +25,7 @@
                 <i class="ti ti-plus"></i>
             </a>
         </div>
+    @endcan
 @endsection
 
 @section('content')
@@ -124,13 +126,18 @@
                                                             </p>
                                                         </div>
                                                         <div class="col-auto text-right">
+                                                            @can('edit interview schedule')
                                                                 <div class="action-btn bg-primary ms-2">
                                                                     <a href="#" data-url="{{ route('meeting.edit',$meeting->id) }}" data-title="{{__('Edit Interview Schedule')}}" data-ajax-popup="true" class="mx-3 btn btn-sm  align-items-center" data-bs-toggle="tooltip" title="{{__('Edit')}}" data-original-title="{{__('Edit')}}"><i class="ti ti-pencil text-white"></i></a>
-                                                                </div>                                                                <div class="action-btn bg-danger ms-2">
+                                                                </div>
+                                                            @endcan
+                                                            @can('delete interview schedule')
+                                                                <div class="action-btn bg-danger ms-2">
                                                                     {!! Form::open(['method' => 'DELETE', 'route' => ['meeting.destroy', $meeting->id],'id'=>'delete-form-'.$meeting->id]) !!}
                                                                     <a href="#" class="mx-3 btn btn-sm  align-items-center bs-pass-para" data-bs-toggle="tooltip" title="{{__('Delete')}}" data-original-title="{{__('Delete')}}" data-confirm="{{__('Are You Sure?').'|'.__('This action can not be undone. Do you want to continue?')}}" data-confirm-yes="document.getElementById('delete-form-{{$meeting->id}}').submit();"><i class="ti ti-trash text-white"></i></a>
                                                                     {!! Form::close() !!}
                                                                 </div>
+                                                            @endcan
                                                         </div>
                                                     </div>
                                                 </div>
